NOVO COMMIT, acabei deixando complexos de mais cheio de regras, então mudei e tirei varias coisas, fiz um main fucional, mas espero amanhar acordar com a cabeça melhor e melhorar o programa

+++ src/main/java/org/primeiroprojetocursooo/projetobancodedados2biblioteca/services/GenericService.java
@@ -13,8 +13,8 @@ public class GenericService<T> {
 
     public void salvar(T entidade) { dao.salvar(entidade); }
     public void atualizar(T entidade) { dao.atualizar(entidade); }
-    public void excluir(Long id) { dao.excluir(id); }
-    public T buscarPorId(Long id) { return dao.buscarPorId(id); }
+    public void excluir(Integer id) { dao.excluir(id); }
+    public T buscarPorId(Integer id) { return dao.buscarPorId(id); }
     public List<T> listar() { return dao.listar(); }
 
     public GenericDAO<T> getDao() {
